+++ src/UAV src/Control/Thrust/MMA.h
@@ -5,16 +5,18 @@
 #include<stddef.h>
 #include"PropThrustControl.h"
 
+//Class for motor mixing algorithm
+
 class MOTOR_MIXING {
     private:
         uint8_t THRUST_RANGE_FIX();
         
     public:
-        uint8_t MOTOR_MIX_ROLL(uint8_t roll);
+        uint8_t MOTOR_MIX_ROLL(double roll);
 
-        uint8_t MOTOR_MIX_YAW(uint8_t yaw);
+        uint8_t MOTOR_MIX_YAW(double yaw);
 
-        uint8_t MOTOR_MIX_PITCH(uint8_t pitch);
+        uint8_t MOTOR_MIX_PITCH(double pitch);
 
 };
